Move MainWindow into ui package
